Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- new parameter "notified" added to delete_subnet_allocator method to delete
  allocator from subnet_allocator dict and not the IndexAllocator from
  ZK (it has already been deleted from the api-serve that received the
  delete request).

- net_(delete/update)_(req/notify) methods modified.
  now, "req" methods delete data from ZooKeeper while
  _subnet_allocators's elements (vnc_db.py) are deleted from "notify"
  methods. _subnet_objs (vnc_addr_mgmt.py) are also deleted from
  "notify" methods

Closes-Bug: #1332645
Closes-Bug: #1365824

+++ src/config/api-server/vnc_db.py
@@ -489,6 +489,7 @@ class VncZkClient(object):
                                 addr_from_start, should_persist,
                                 start_subnet, size, alloc_unit):
         # TODO handle subnet resizing change, ignore for now
+
         if subnet not in self._subnet_allocators:
             if addr_from_start is None:
                 addr_from_start = False
@@ -501,10 +502,12 @@ class VncZkClient(object):
                 max_alloc=self._MAX_SUBNET_ADDR_ALLOC/alloc_unit)
     # end create_subnet_allocator
 
-    def delete_subnet_allocator(self, subnet):
-        self._subnet_allocators.pop(subnet, None)
-        IndexAllocator.delete_all(self._zk_client,
-                                  self._subnet_path+'/'+subnet+'/')
+    def delete_subnet_allocator(self, subnet_name, notified=True):
+        if notified:
+            self._subnet_allocators.pop(subnet_name, None)
+        else:
+            IndexAllocator.delete_all(self._zk_client,
+                                   self._subnet_path+'/'+subnet_name+'/')
     # end delete_subnet_allocator
 
     def _get_subnet_allocator(self, subnet):
@@ -1310,8 +1313,8 @@ class VncDbClient(object):
                                should_persist, start_subnet, size, alloc_unit)
     # end subnet_create_allocator
 
-    def subnet_delete_allocator(self, subnet):
-        return self._zk_db.delete_subnet_allocator(subnet)
+    def subnet_delete_allocator(self, subnet, notified=True):
+        return self._zk_db.delete_subnet_allocator(subnet, notified)
     # end subnet_delete_allocator
 
     def uuid_vnlist(self):
